Removed the three lines image and replaced with just "Menu"
When you run complilations in Mono you will sometimes run into issues where the images
for libraries don't get into the bundle. That's terrible. To prevent that, the images should always be
included from the main project, not the libraries.

The CreateMenuButton function can be overrided and a UIBarButtonItem with an image can be returned
if the user wants this to be customizable.